Fix Windows portability and clean up stale TODO items
Add build-tag stubs so the provider package compiles on Windows by
extracting PTY-specific startup probe functions into platform files.
Make the server status message platform-aware (TCP localhost on Windows
vs unix socket elsewhere). Switch pki_test to t.TempDir() for automatic
cleanup. Remove stale TODO items that were already fixed in prior PRs
and move cert lifecycle items to GitHub Issues #225 and #224.
